🚚 files: 공통이 아닌 컴포넌트를 페이지 내로 이동
