--- src/components/themeset/ThemeSet.jsx
@@ -1,12 +1,0 @@
-import ThemeMenu from '@components/themeset/ThemeMenu';
-import { ThemeSetting } from '@components/themeset/ThemeSet.style';
-
-function ThemeSet() {
-  return (
-    <ThemeSetting>
-      <ThemeMenu />
-    </ThemeSetting>
-  );
-}
-
-export default ThemeSet;
