--- src/components/timeset/TimeSet.jsx
@@ -1,12 +1,0 @@
-import TimeMenu from '@components/timeset/TimeMenu';
-import { TimeSetting } from '@components/timeset/TimeSet.style';
-
-function TimeSet() {
-  return (
-    <TimeSetting>
-      <TimeMenu />
-    </TimeSetting>
-  );
-}
-
-export default TimeSet;
