+++ src/pages/meditation/themeset/ThemeSet.jsx
@@ -1,0 +1,12 @@
+import ThemeMenu from '@pages/meditation/themeset/ThemeMenu';
+import { ThemeSetting } from '@pages/meditation/themeset/ThemeSet.style';
+
+function ThemeSet() {
+  return (
+    <ThemeSetting>
+      <ThemeMenu />
+    </ThemeSetting>
+  );
+}
+
+export default ThemeSet;
