+++ src/pages/meditation/timeset/TimeSet.jsx
@@ -1,0 +1,12 @@
+import TimeMenu from '@pages/meditation/timeset/TimeMenu';
+import { TimeSetting } from '@pages/meditation/timeset/TimeSet.style';
+
+function TimeSet() {
+  return (
+    <TimeSetting>
+      <TimeMenu />
+    </TimeSetting>
+  );
+}
+
+export default TimeSet;
